addresses four issues: 1. **README.md** - documents the reset-on-target-change behavior. 2. **`updateTable()`** - resolves `_initialSortColumn` into a local `sortColumn`, warns via `console.warn()` and falls back to `'model_id'` if the column isn't present in the current table, then uses `sortColumn` in the DataTables `order` config. 3. **`showDialog()`** - added a comment explaining some Bootstrap-modal tradeoffs and why `window.alert()` is preferred for this developer-facing, rare error path. 4. **`src/validation.js`** - new file containing `_validateOptions` as a standalone exported function, following the predtimechart pattern.
